+++ CHANGELOG.md
@@ -64,3 +64,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Github CI
 - VUE_APP_PUBLIC_PATH
 - Nginx conf
+- Config Loader